requirements.txt
fastapi
uvicorn
psycopg2
pymupdf
langchain
faiss-cpu
sentence-transformers
requests
python-dotenv
